requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
openai